227.2 Building a List with Redux Data(Section 17: Integrating React with Redux)
In our original mockup showed it kind of being a little bit constrained in width and it also should

have kind of a margin on the left hand side as well.

So to take care of that we're going to set up a grid system just like we did back on that video.

Now remember the grid system a semantic UI is not the same thing as the CSSA grid system.

So back inside of my app component is the app component.

We're just going to add a couple of elements in a couple of class names to this thing to give our app

a little bit more structure when it comes to the layout of it.

So on this top level div I'm going to give it a class name of UI container grid Eildon Place a div inside

there with the class name of UI row and I'm going to close off the div on the other side of the song

list and I'm going to add in one more nested if you're with the class name of column 8 wide

and again I'll close it off on the other side the song list.

So we've now got three nested divs right here each of which have a class name.

We've got content.

Can you give me your container grid you I row and column 8 wide.

So let's say I'll save this we'll flip back over and that looks way more reasonable.

So it looks a lot more like what we had initially been imagining inside of our Mocca.

OK.

So this looks pretty good.

Let's take a pause right here.

When we come back the next section we're going to make sure that any time a user clicks on the select

button we will somehow select that song and then attempt to show information about it on the right hand

side of her app.

So quick pause and I'll see you in just a minute.