228.1 Calling Action Creators from Components(Section 17: Integrating React with Redux)
Unless section we're able to render out our list of songs on the screen.

So now we need to make sure that any time that a user clicks on one of these select buttons we somehow

try to select that song so to speak.

Now when we select a song what we really want to do is update the data where the state of our redux

store.

Remember any time that we want to update our data with redux we have to call an action creator.

Now we've already put together an action creator inside of our actions index not just file that will

attempt to select a song.

So we call this action create or select song.

If we pass a song object to it it will return in action with it type of song selected.

And our payload of song.

This action should then end up inside of our selected song reducer which is inside of our reducers index

file.

So this reducer is going to look at the action type if it is equal to song selected then it's going

to return that particular song.

Otherwise if it is a action with a different type then it's going to just return whatever the currently

selected song is.

All right.

So all we really have to do now is figure out how to call that action creator.

As a quick reminder I told you in passing a couple of videos ago that the Kinect function is going to

be used not only to get data out of our store but the connect function can also be used to kind of get

action creators correctly into our song list.

So we're going to import our action creator that we created number that function that we just looked

at two seconds ago and we're going to essentially wire it up to this connect helper.

All right so back inside my song list at the very top I'm going to import select song from actions.

Now remember when we created select song the action creator we export it as a named export from this

file.

That is why we are using the curly braces right here.

If we had exported it using the default keyword we would not use the curly braces.

All right so now that we've got our action creator this is the plain function by itself.

We're going to pass it into that connect function at the bottom.

So I'm going to go down to the bottom and then as a second argument to the connect function I'm going

to pass in an object's this object is going to have a key of select song and a value of the action creator

that we just imported.