228.2 Calling Action Creators from Components(Section 17: Integrating React with Redux)
Notice now that the key and the value inside this object are identical.

So we can use a little bit of yes 20:15 syntax to shorten this up and reduce down to just say select

song like so still inside of a set of curly braces.

All right so there's the line in total now that we are passing in a second argument here the connect

function is going to take that select song action creator and pass it into our component as a prop to

again inside the render method a log.

This prompts.

Now when I save this you'll notice that my code formatter kicked in and did a little bit of a reactor

right there.

It still totally equivalent to what you just saw a moment ago.

All right.

So I'm going to save this file and we'll flip back over.

It's now inside of my console log right here remember.

This is my prop's object.

I see that I have a function here called select song.

So that's our action creator.

And if we call it it's going to automatically take the action that gets returned and send it in to read

ox's dispatch function.

Now when I say that last sentence right there I said take the action that gets returned and send it

into the dispatch function.

I'm kind of glossing over a tremendous amount of complexity.

So what is the complexity there.