231.1. Conditional Rendering(Section 17: Integrating React with Redux)
In the last section we were able to print out the currently selected song from our song detail component

by hooking up the connect function to it.

So now we need to somehow customize what the song detail is going to render out based upon the song

inside that props object.

Now before we do one quick thing I want mention though to set our props object has a key of song and

the song has a value of the actual song object right there.

We see the key of song right there only because that is what we had returned from our map state to prop's

function.

So right here is an object that says song that key right there is where this entire object I should

say is what is going to eventually show up as our props object inside the component.

So if we wanted to we could rename that key right there to something like my selected song.

And then if I saved this and foot back over you'll see that the props object has now changed and it

now says my selected song.

And when I actually select a song yet there we go.

All right.

So if you ever want to customize the key that shows up inside your props object.

You do it right here inside a map state to props.

But for us having simply a song like so is totally appropriate for what we are doing.

OK so now back inside of our component itself I'm going to move the console log now inside of here.

I don't really want to have to refer to prop stot song title.

A whole bunch.

And then like you know props start song duration over and over again.

So instead of receiving the entire props object I'm going to structure out just the property we care

about just the song property like the it's now inside of here.

I should be able to reference song title and see the title of the song printed on the screen.

So if I now flip back over I'm going to see an error message.

All right.

So cannot read property title of No.

Why are we seeing this.