234.1 Initial App Setup{yarn add redux-thunk --save}(Section 18: Async Actions with Redux Thunk)
Unless we generated our brand new project I'm not going to change into that directory that was created.

I'm going to start at my code editor and site there and then we're also going to install a couple of

additional dependencies inside of our project right now rather than having to come back and install

these dependencies later on.

So inside of my blog project directory I'm going to do an NPM install dash dash save and we're going

to get redux reactor's ducks X-posts.

And then one other new dependency that we're going to talk about quite a bit inside this application

called redux Dasch think like so redux Dasch thunk.

That's the new one.

OK.

So I'm going to run this.

And while that's installing.

Let's have a quick reminder about what all these dependencies do including this new one called redux

thunk.

All right.

So first off redux.

Yup.

You know what that is.

We've seen good examples of that already.

It's the actual redux library itself.

Next up reactor redux.

We just worked on an application that made use of this library.

Remember this is the integration layer between the react and redux libraries because they were not originally

designed to work with each other.

Right out of the box we've made use of axially before member.

We make use of ASIO's any time that we want to make network requests.

Ex-U.S. is not strictly required inside of these Riak projects.

We could technically use that fetch API or the fecche function.

But remember we had said that there's a lot of kind of manual hand-holding that you have to do with

the fetch function.

And so it's usually just a lot easier to make use of XS.

And then finally this new dependency called redux thunk.

Now this is a dependency that we refer to as a middleware the term middleware applies to the redux library

middleware are essentially functions that are going to slightly change the behavior of our redux store.

So they will add in new capabilities or new features to the redux side of your application.

In the case of redux thunk this is a middleware that is going to help us make network requests from

the read up side of rap in the vast vast majority of redux applications you are ever going to work on.

You you're going to make use of a middleware like redux thunk.

Maybe not exactly redux stunk but something very much like it.

Redux thunk is extremely popular and there's only a handful of other alternatives that make it easy

for you to do network requests from the redux side of your application.

So again if not redux thunk it will be something very similar to it.

All right so let's flip back over to our terminal.

It looks like everything has installed successfully.

So I'm not going to start up my application with NPM start.

And after a second or two we should see the application.

OK it started up but I already deleted some boilerplate inside the SIRC folder.

So you're probably going to see that default loading screen appear right now.

So let's flip back over to our code editor.

We're going to go ahead and start doing some cleanup inside that sarcy directory.

So as usual I'm going to delete everything inside my sarcy folder and we're going to replace all that

boilerplate automatically generated code with some usual boilerplate that we're going to use inside

of just about every react and redux application that you and I work on.

All right.